requirements.txt
crewai>=0.100.0
crewai-tools>=0.20.0
tavily-python>=0.5.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
